Fix PlayerMovement Final
Finally managed to fix player movement by making the speed consistent as well as the direction the player is sent in.

Before, if the player was not centred at 0,0 (e.g. they were on (-1, -2) and I clicked on (-1, -4) to send them upwards, they would be sent diagonally upwards to (1,4) instead.